Fix module compilation warnings
and change amx_Address to amx_GetAddr

+++ amx-deps/src/CFunctions.cpp
@@ -261,7 +261,7 @@ int CFunctions::amxCall(lua_State *luaVM) {
 				std::string str = lua_tostring(luaVM, i);
 				if(str.find(".")!=std::string::npos)
 				{
-					float fval = lua_tonumber(luaVM, i);
+					float fval = (float)lua_tonumber(luaVM, i);
 					cell val = *(cell*)&fval;
 					amx_Push(amx, val);
 				}
